backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
asyncpg
psycopg2-binary
alembic
passlib[bcrypt]
python-jose
python-multipart
python-barcode
pillow
multipart